feat: comprehensive project maintenance and documentation restructure
This commit addresses multiple GitHub issues for better project organization,
code maintainability, and documentation structure.

##  Folder Structure Improvements
- Rename cmd/  examples/ to follow Go project layout standards
- All examples relocated with preserved functionality
- Validated build and runtime compatibility

##  Code Cleanup
- Remove deprecated AddStandardVariables() helper method from FlightDataManager
- Remove deprecated AddWritableStandardVariables() helper method
- Update affected examples to use explicit variable definitions
- Maintain backward compatibility for core API

##  Documentation Restructure
- Create comprehensive docs/ hierarchy with clear navigation
- Add detailed API reference documentation (Client, FlightDataManager, Variables)
- Create getting-started guide with installation and quick start
- Add advanced topics section with troubleshooting guide
- Add example documentation with usage patterns
- Update main README.md to be concise with structured links

##  New Documentation Structure
docs/
 README.md                    # Documentation navigation
 getting-started.md          # Installation & quick start
 api/                        # Complete API reference
    client.md              # Core SimConnect client
    flight-data-manager.md # High-level data management
    variables.md           # 200+ SimConnect variables
 examples/                   # Example documentation
    README.md              # Example guides and patterns
 advanced/                   # Advanced topics
     README.md              # Advanced topics overview
     troubleshooting.md     # Common issues & solutions

##  Validation
- All examples build and run correctly from new location
- Core functionality validated with live MSFS 2024 connection
- Camera control test confirms SetData operations working
- Package builds without errors after all changes

##  Impact
- Better project organization following Go standards
- Cleaner API without deprecated helper methods
- Comprehensive documentation for better developer experience
- Future-proof structure for continued development

Closes #9
Closes #11
Closes #12
