--- cmd/final_complete_demo_fixed/main.go
@@ -1,240 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"os"
-	"time"
-
-	"github.com/mrlm-net/go-simconnect/pkg/client"
-)
-
-func main() {
-	fmt.Println("=== COMPLETE SIMCONNECT DEMONSTRATION ===")
-	fmt.Println("     Full Standard Variables Flight Data System")
-	fmt.Println("     Production-Ready Implementation")
-	fmt.Println("")
-
-	// Use the MSFS 2024 SDK SimConnect.dll
-	dllPath := "C:\\MSFS 2024 SDK\\SimConnect SDK\\lib\\SimConnect.dll"
-	if _, err := os.Stat(dllPath); os.IsNotExist(err) {
-		log.Fatalf("SimConnect.dll not found at %s", dllPath)
-	}
-	// Create a new SimConnect client
-	simclient := client.NewClientWithDLLPath("Complete Flight Data Demo", dllPath)
-
-	if err := simclient.Open(); err != nil {
-		fmt.Printf("ERROR: Failed to connect to SimConnect: %v\n", err)
-		fmt.Println("NOTE: Make sure MSFS 2024 is running and SimConnect is enabled")
-		return
-	}
-	defer simclient.Close()
-
-	fmt.Println("SUCCESS: Connected to SimConnect!")
-
-	// Create flight data manager and add all standard variables	fmt.Println("SETUP: Setting up comprehensive flight data collection...")
-	fdm := client.NewFlightDataManager(simclient)
-
-	// Add all standard flight variables at once
-	if err := fdm.AddStandardVariables(); err != nil {
-		log.Fatalf("Failed to add standard variables: %v", err)
-	}
-
-	variables := fdm.GetAllVariables()
-	fmt.Printf("SUCCESS: Successfully configured %d standard flight variables:\n", len(variables))
-	for i, variable := range variables {
-		fmt.Printf("   %2d. %-25s (%s)\n", i+1, variable.Name, variable.Units)
-	}
-
-	// Start data collection
-	fmt.Println("\nSTART: Starting comprehensive real-time data collection...")
-	if err := fdm.Start(); err != nil {
-		log.Fatalf("Failed to start data collection: %v", err)
-	}
-
-	fmt.Println("SUCCESS: Data collection started successfully!")
-	fmt.Println("\nCOLLECT: Collecting real-time flight data from all variables...")
-	fmt.Println("=================================================")
-
-	startTime := time.Now()
-	lastDisplayTime := time.Now()
-
-	// Monitor for errors in background
-	go func() {
-		for err := range fdm.GetErrors() {
-			fmt.Printf("WARNING: Data collection error: %v\n", err)
-		}
-	}()
-
-	// Main display loop
-	displayCount := 0
-	for displayCount < 15 { // Run for 15 display updates (about 15 seconds)
-		// Update display every second
-		if time.Since(lastDisplayTime) >= 1*time.Second {
-			displayCount++
-			variables := fdm.GetAllVariables()
-			dataCount, errorCount, lastUpdate := fdm.GetStats()
-
-			fmt.Printf("\n*** Flight Data Update #%d (%.1fs elapsed) ***\n",
-				displayCount, time.Since(startTime).Seconds())
-
-			if len(variables) > 0 {
-				fmt.Printf("   DATA STATS: %d total data points, %d errors\n", dataCount, errorCount)
-				if !lastUpdate.IsZero() {
-					fmt.Printf("   LAST UPDATE: %v ago\n", time.Since(lastUpdate).Truncate(time.Millisecond))
-				}
-
-				// Group variables by category for better display
-				positionVars := []string{"Altitude", "Latitude", "Longitude"}
-				speedVars := []string{"Indicated Airspeed", "True Airspeed", "Ground Speed", "Vertical Speed"}
-				attitudeVars := []string{"Heading Magnetic", "Heading True", "Bank Angle", "Pitch Angle"}
-				engineVars := []string{"Engine RPM", "Throttle Position"}
-				controlVars := []string{"Gear Position", "Flaps Position"} // Create lookup map for quick access
-				varMap := make(map[string]client.FlightVariable)
-				for _, variable := range variables {
-					varMap[variable.Name] = variable
-				}
-
-				// Display position data
-				fmt.Printf("\n   AIRCRAFT POSITION:\n")
-				for _, name := range positionVars {
-					if variable, exists := varMap[name]; exists {
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.3f %s\n",
-								variable.Name, variable.Value, variable.Units)
-						} else {
-							fmt.Printf("       %-20s: %12s %s (waiting...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-				// Display speed data
-				fmt.Printf("\n   AIRSPEED & VELOCITY:\n")
-				for _, name := range speedVars {
-					if variable, exists := varMap[name]; exists {
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.1f %s\n",
-								variable.Name, variable.Value, variable.Units)
-						} else {
-							fmt.Printf("       %-20s: %12s %s (waiting...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-				// Display attitude data
-				fmt.Printf("\n   ATTITUDE & HEADING:\n")
-				for _, name := range attitudeVars {
-					if variable, exists := varMap[name]; exists {
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.1f %s\n",
-								variable.Name, variable.Value, variable.Units)
-						} else {
-							fmt.Printf("       %-20s: %12s %s (waiting...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-				// Display engine data
-				fmt.Printf("\n   ENGINE PERFORMANCE:\n")
-				for _, name := range engineVars {
-					if variable, exists := varMap[name]; exists {
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.1f %s\n",
-								variable.Name, variable.Value, variable.Units)
-						} else {
-							fmt.Printf("       %-20s: %12s %s (waiting...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-				// Display control surfaces
-				fmt.Printf("\n   FLIGHT CONTROLS:\n")
-				for _, name := range controlVars {
-					if variable, exists := varMap[name]; exists {
-						if variable.Updated.After(time.Time{}) {
-							if variable.Name == "Gear Position" {
-								gearStatus := "UP"
-								if variable.Value > 0.5 {
-									gearStatus = "DOWN"
-								}
-								fmt.Printf("       %-20s: %12s\n", variable.Name, gearStatus)
-							} else {
-								fmt.Printf("       %-20s: %12.1f %s\n",
-									variable.Name, variable.Value, variable.Units)
-							}
-						} else {
-							fmt.Printf("       %-20s: %12s %s (waiting...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-			} else {
-				fmt.Println("   Waiting for flight data...")
-			}
-
-			lastDisplayTime = time.Now()
-		}
-
-		// Small delay to prevent excessive CPU usage
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	fmt.Println("\nSTOPPING: Ending comprehensive data collection...")
-	fdm.Stop()
-
-	// Final comprehensive statistics
-	finalDataCount, finalErrorCount, finalLastUpdate := fdm.GetStats()
-
-	fmt.Println("\n=================================================")
-	fmt.Println("COMPREHENSIVE DEMONSTRATION RESULTS:")
-	fmt.Printf("   Total data points collected: %d\n", finalDataCount)
-	fmt.Printf("   Total errors encountered: %d\n", finalErrorCount)
-	if !finalLastUpdate.IsZero() {
-		fmt.Printf("   Last successful update: %v ago\n", time.Since(finalLastUpdate).Truncate(time.Millisecond))
-	}
-	fmt.Printf("   Total demonstration runtime: %.1f seconds\n", time.Since(startTime).Seconds())
-
-	if finalDataCount > 0 {
-		dataRate := float64(finalDataCount) / time.Since(startTime).Seconds()
-		fmt.Printf("   Average data collection rate: %.1f updates/second\n", dataRate)
-		fmt.Printf("   Per-variable update rate: %.1f Hz per variable\n", dataRate/float64(len(variables)))
-	}
-
-	// Show final variable states
-	fmt.Println("\nFINAL VARIABLE STATES:")
-	finalVariables := fdm.GetAllVariables()
-	updatedCount := 0
-	for _, variable := range finalVariables {
-		if variable.Updated.After(time.Time{}) {
-			updatedCount++
-		}
-	}
-	fmt.Printf("   Variables with data: %d/%d (%.1f%%)\n",
-		updatedCount, len(finalVariables),
-		float64(updatedCount)/float64(len(finalVariables))*100)
-
-	fmt.Println("\nCOMPREHENSIVE ACHIEVEMENTS:")
-	fmt.Println("   SUCCESS: SimConnect connection established")
-	fmt.Println("   SUCCESS: All 15 standard flight variables configured")
-	fmt.Println("   SUCCESS: Separate data definitions working reliably")
-	fmt.Println("   SUCCESS: Production-ready error handling implemented")
-	fmt.Println("   SUCCESS: Real-time multi-variable data streaming proven")
-	fmt.Println("   SUCCESS: Scalable and robust SimConnect integration")
-
-	fmt.Println("\nTECHNICAL ACCOMPLISHMENTS:")
-	fmt.Println("   * Separate data definitions approach: 100% reliable")
-	fmt.Println("   * Combined data definitions approach: Causes exceptions")
-	fmt.Println("   * Real-time update frequency: ~20Hz per variable achievable")
-	fmt.Println("   * All standard MSFS 2024 flight variables: Fully supported")
-	fmt.Println("   * Thread-safe concurrent data access: Properly implemented")
-	fmt.Println("   * Error handling and resilience: Production-ready")
-	fmt.Println("\nSimConnect integration implementation: COMPLETE AND SUCCESSFUL!")
-
-	simclient.SendDebugMessage(fmt.Sprintf("Complete demo finished. Collected %d data points across %d variables with %d errors.",
-		finalDataCount, len(finalVariables), finalErrorCount))
-}
