--- cmd/final_production_demo/main.go
@@ -1,231 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"os"
-	"time"
-
-	"github.com/mrlm-net/go-simconnect/pkg/client"
-)
-
-func main() {
-	fmt.Println("🛩️  === FINAL PRODUCTION SIMCONNECT DEMONSTRATION ===")
-	fmt.Println("     Real-time Flight Data Collection System")
-	fmt.Println("     Using Optimized Separate Data Definitions Approach")
-	fmt.Println("")
-
-	// Use the MSFS 2024 SDK SimConnect.dll
-	dllPath := "C:\\MSFS 2024 SDK\\SimConnect SDK\\lib\\SimConnect.dll"
-	if _, err := os.Stat(dllPath); os.IsNotExist(err) {
-		log.Fatalf("SimConnect.dll not found at %s", dllPath)
-	}
-	// Create a new SimConnect client
-	simclient := client.NewClientWithDLLPath("Production Flight Data Demo", dllPath)
-
-	if err := simclient.Open(); err != nil {
-		fmt.Printf("❌ Failed to connect to SimConnect: %v\n", err)
-		fmt.Println("💡 Make sure MSFS 2024 is running and SimConnect is enabled")
-		return
-	}
-	defer simclient.Close()
-
-	fmt.Println("✅ Successfully connected to SimConnect!")
-
-	// Create flight data manager
-	fdm := client.NewFlightDataManager(simclient)
-
-	// Add comprehensive set of flight variables
-	fmt.Println("📊 Setting up flight data variables...")
-
-	// Core navigation data
-	if err := fdm.AddVariable("Altitude", "Plane Altitude", "feet"); err != nil {
-		log.Fatalf("Failed to add Altitude: %v", err)
-	}
-	if err := fdm.AddVariable("Indicated Airspeed", "Airspeed Indicated", "knots"); err != nil {
-		log.Fatalf("Failed to add Indicated Airspeed: %v", err)
-	}
-	if err := fdm.AddVariable("True Airspeed", "Airspeed True", "knots"); err != nil {
-		log.Fatalf("Failed to add True Airspeed: %v", err)
-	}
-	if err := fdm.AddVariable("Ground Speed", "Ground Velocity", "knots"); err != nil {
-		log.Fatalf("Failed to add Ground Speed: %v", err)
-	}
-
-	// Position data
-	if err := fdm.AddVariable("Latitude", "Plane Latitude", "degrees"); err != nil {
-		log.Fatalf("Failed to add Latitude: %v", err)
-	}
-	if err := fdm.AddVariable("Longitude", "Plane Longitude", "degrees"); err != nil {
-		log.Fatalf("Failed to add Longitude: %v", err)
-	}
-
-	// Heading and attitude
-	if err := fdm.AddVariable("Heading Magnetic", "Plane Heading Degrees Magnetic", "degrees"); err != nil {
-		log.Fatalf("Failed to add Heading Magnetic: %v", err)
-	}
-	if err := fdm.AddVariable("Bank Angle", "Plane Bank Degrees", "degrees"); err != nil {
-		log.Fatalf("Failed to add Bank Angle: %v", err)
-	}
-	if err := fdm.AddVariable("Pitch Angle", "Plane Pitch Degrees", "degrees"); err != nil {
-		log.Fatalf("Failed to add Pitch Angle: %v", err)
-	}
-
-	// Vertical navigation
-	if err := fdm.AddVariable("Vertical Speed", "Vertical Speed", "feet per minute"); err != nil {
-		log.Fatalf("Failed to add Vertical Speed: %v", err)
-	}
-
-	// Engine and control data
-	if err := fdm.AddVariable("Engine RPM", "General Eng RPM:1", "rpm"); err != nil {
-		log.Fatalf("Failed to add Engine RPM: %v", err)
-	}
-	if err := fdm.AddVariable("Throttle Position", "General Eng Throttle Lever Position:1", "percent"); err != nil {
-		log.Fatalf("Failed to add Throttle Position: %v", err)
-	}
-
-	fmt.Printf("✅ Successfully configured %d flight variables\n", 12)
-
-	// Start data collection
-	fmt.Println("\n🚁 Starting real-time data collection...")
-	if err := fdm.Start(); err != nil {
-		log.Fatalf("Failed to start data collection: %v", err)
-	}
-
-	fmt.Println("✅ Data collection started successfully!")
-	fmt.Println("\n📡 Collecting real-time flight data...")
-	fmt.Println("=" + "============================================" + "=")
-
-	startTime := time.Now()
-	lastDisplayTime := time.Now()
-
-	// Monitor for errors in background
-	go func() {
-		for err := range fdm.GetErrors() {
-			fmt.Printf("⚠️  Data collection warning: %v\n", err)
-		}
-	}()
-
-	// Main display loop
-	displayCount := 0
-	for displayCount < 30 { // Run for 30 display updates (about 30 seconds)
-		// Update display every second
-		if time.Since(lastDisplayTime) >= 1*time.Second {
-			displayCount++
-			variables := fdm.GetAllVariables()
-			dataCount, errorCount, lastUpdate := fdm.GetStats()
-
-			// Clear screen for better display (optional)
-			fmt.Printf("\n📦 Flight Data Update #%d (%.1fs elapsed)\n",
-				displayCount, time.Since(startTime).Seconds())
-
-			if len(variables) > 0 {
-				fmt.Printf("   📊 Stats: %d data points collected, %d errors, last update: %v ago\n",
-					dataCount, errorCount, time.Since(lastUpdate).Truncate(time.Millisecond))
-
-				fmt.Printf("\n   🛩️  AIRCRAFT POSITION & NAVIGATION:\n")
-				for _, variable := range variables {
-					switch variable.Name {
-					case "Altitude", "Latitude", "Longitude":
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.3f %-15s (updated %v ago)\n",
-								variable.Name, variable.Value, variable.Units,
-								time.Since(variable.Updated).Truncate(time.Millisecond))
-						} else {
-							fmt.Printf("       %-20s: %12s %-15s (waiting for data...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-				fmt.Printf("\n   💨 AIRSPEED & PERFORMANCE:\n")
-				for _, variable := range variables {
-					switch variable.Name {
-					case "Indicated Airspeed", "True Airspeed", "Ground Speed", "Vertical Speed":
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.1f %-15s (updated %v ago)\n",
-								variable.Name, variable.Value, variable.Units,
-								time.Since(variable.Updated).Truncate(time.Millisecond))
-						} else {
-							fmt.Printf("       %-20s: %12s %-15s (waiting for data...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-				fmt.Printf("\n   📐 ATTITUDE & HEADING:\n")
-				for _, variable := range variables {
-					switch variable.Name {
-					case "Heading Magnetic", "Bank Angle", "Pitch Angle":
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.1f %-15s (updated %v ago)\n",
-								variable.Name, variable.Value, variable.Units,
-								time.Since(variable.Updated).Truncate(time.Millisecond))
-						} else {
-							fmt.Printf("       %-20s: %12s %-15s (waiting for data...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-				fmt.Printf("\n   ⚙️  ENGINE & CONTROLS:\n")
-				for _, variable := range variables {
-					switch variable.Name {
-					case "Engine RPM", "Throttle Position":
-						if variable.Updated.After(time.Time{}) {
-							fmt.Printf("       %-20s: %12.1f %-15s (updated %v ago)\n",
-								variable.Name, variable.Value, variable.Units,
-								time.Since(variable.Updated).Truncate(time.Millisecond))
-						} else {
-							fmt.Printf("       %-20s: %12s %-15s (waiting for data...)\n",
-								variable.Name, "---", variable.Units)
-						}
-					}
-				}
-
-			} else {
-				fmt.Println("   ⏳ Waiting for flight data...")
-			}
-
-			lastDisplayTime = time.Now()
-		}
-
-		// Small delay to prevent excessive CPU usage
-		time.Sleep(100 * time.Millisecond)
-	}
-
-	fmt.Println("\n🏁 Stopping data collection...")
-	fdm.Stop()
-
-	// Final statistics
-	finalDataCount, finalErrorCount, finalLastUpdate := fdm.GetStats()
-
-	fmt.Println("\n" + "=" + "============================================" + "=")
-	fmt.Println("📈 FINAL RESULTS:")
-	fmt.Printf("   ✅ Total data points collected: %d\n", finalDataCount)
-	fmt.Printf("   ⚠️  Total errors encountered: %d\n", finalErrorCount)
-	fmt.Printf("   🕒 Last successful update: %v ago\n", time.Since(finalLastUpdate).Truncate(time.Millisecond))
-	fmt.Printf("   ⏱️  Total runtime: %.1f seconds\n", time.Since(startTime).Seconds())
-
-	if finalDataCount > 0 {
-		fmt.Printf("   📊 Average data rate: %.1f updates/second\n", float64(finalDataCount)/time.Since(startTime).Seconds())
-	}
-
-	fmt.Println("\n💡 KEY ACHIEVEMENTS:")
-	fmt.Println("   ✅ Successfully established SimConnect connection")
-	fmt.Println("   ✅ Configured separate data definitions for each variable")
-	fmt.Println("   ✅ Collected real-time flight simulation data")
-	fmt.Println("   ✅ Demonstrated production-ready error handling")
-	fmt.Println("   ✅ Proved scalable multi-variable data collection")
-
-	fmt.Println("\n🎯 TECHNICAL INSIGHTS:")
-	fmt.Println("   • Separate data definitions approach works reliably")
-	fmt.Println("   • Combined data definitions cause SimConnect exceptions")
-	fmt.Println("   • Real-time update frequency: ~20Hz achievable")
-	fmt.Println("   • All standard flight variables are supported")
-	fmt.Println("   • Thread-safe data access with proper synchronization")
-
-	fmt.Println("\n✅ Production SimConnect implementation demonstration completed successfully!")
-
-	simclient.SendDebugMessage(fmt.Sprintf("Production demo completed. Collected %d data points with %d errors.", finalDataCount, finalErrorCount))
-}
