+++ docs/advanced/README.md
@@ -1,0 +1,74 @@
+# Advanced Topics
+
+This section covers advanced usage patterns, optimization techniques, and architectural considerations for building robust SimConnect applications with go-simconnect.
+
+## Table of Contents
+
+- [Performance Optimization](performance.md) - Techniques for high-performance applications
+- [Troubleshooting Guide](troubleshooting.md) - Common issues and solutions
+- [Architecture Patterns](architecture.md) - Design patterns for complex applications
+- [Production Deployment](deployment.md) - Best practices for production environments
+- [Testing Strategies](testing.md) - Unit testing and integration testing approaches
+- [Error Handling](error-handling.md) - Robust error handling patterns
+- [Memory Management](memory-management.md) - Optimizing memory usage and preventing leaks
+
+## Performance Optimization
+
+Learn how to optimize your SimConnect applications for:
+- High-frequency data collection
+- Minimal CPU and memory usage
+- Low-latency command execution
+- Efficient network utilization
+
+[Read Performance Guide →](performance.md)
+
+## Troubleshooting
+
+Common issues and their solutions:
+- Connection problems
+- Data synchronization issues
+- Performance bottlenecks
+- Memory leaks and resource management
+
+[Read Troubleshooting Guide →](troubleshooting.md)
+
+## Architecture Patterns
+
+Design patterns for building maintainable and scalable SimConnect applications:
+- Model-View-Controller (MVC) patterns
+- Event-driven architectures
+- Microservice integration
+- Plugin systems
+
+[Read Architecture Guide →](architecture.md)
+
+## Production Deployment
+
+Best practices for deploying SimConnect applications in production:
+- Configuration management
+- Monitoring and alerting
+- Health checks and recovery
+- Security considerations
+
+[Read Deployment Guide →](deployment.md)
+
+## Testing Strategies
+
+Comprehensive testing approaches:
+- Unit testing with mocks
+- Integration testing with simulator
+- Performance testing
+- Continuous integration setup
+
+[Read Testing Guide →](testing.md)
+
+## Contributing to Advanced Documentation
+
+We welcome contributions to improve this documentation:
+
+1. **Share Your Experience** - Document patterns you've found useful
+2. **Report Issues** - Help us identify gaps or outdated information
+3. **Provide Examples** - Add real-world examples of advanced usage
+4. **Improve Clarity** - Help make complex topics more accessible
+
+See our [Contributing Guide](../../CONTRIBUTING.md) for details on how to contribute.
