+++ examples/camera_test/main.go
