+++ examples/web_dashboard/static/app.js
