+++ pkg/client/flight_data.go
@@ -78,71 +78,6 @@ func (fdm *FlightDataManager) AddVariableWithWritable(name, simVar, units string
 	fdm.variables = append(fdm.variables, variable)
 	fdm.definitions = append(fdm.definitions, defineID)
 	fdm.requests = append(fdm.requests, requestID)
-
-	return nil
-}
-
-// AddStandardVariables adds a set of commonly used flight simulation variables
-func (fdm *FlightDataManager) AddStandardVariables() error {
-	standardVars := []struct {
-		name     string
-		simVar   string
-		units    string
-		writable bool
-	}{
-		{"Altitude", "Plane Altitude", "feet", false},                                   // Read-only
-		{"Indicated Airspeed", "Airspeed Indicated", "knots", false},                    // Read-only
-		{"True Airspeed", "Airspeed True", "knots", false},                              // Read-only
-		{"Ground Speed", "Ground Velocity", "knots", false},                             // Read-only
-		{"Latitude", "Plane Latitude", "degrees", false},                                // Read-only
-		{"Longitude", "Plane Longitude", "degrees", false},                              // Read-only
-		{"Heading Magnetic", "Plane Heading Degrees Magnetic", "degrees", false},        // Read-only
-		{"Heading True", "Plane Heading Degrees True", "degrees", false},                // Read-only
-		{"Bank Angle", "Plane Bank Degrees", "degrees", false},                          // Read-only
-		{"Pitch Angle", "Plane Pitch Degrees", "degrees", false},                        // Read-only
-		{"Vertical Speed", "Vertical Speed", "feet per minute", false},                  // Read-only
-		{"Engine RPM", "General Eng RPM:1", "rpm", false},                               // Read-only
-		{"Throttle Position", "General Eng Throttle Lever Position:1", "percent", true}, // Writable
-		{"Gear Position", "Gear Handle Position", "bool", true},                         // Writable
-		{"Flaps Position", "Flaps Handle Percent", "percent", true},                     // Writable
-	}
-
-	for _, variable := range standardVars {
-		if err := fdm.AddVariableWithWritable(variable.name, variable.simVar, variable.units, variable.writable); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// AddWritableStandardVariables adds commonly writable flight simulation variables
-func (fdm *FlightDataManager) AddWritableStandardVariables() error {
-	writableVars := []struct {
-		name   string
-		simVar string
-		units  string
-	}{
-		{"Throttle Position", "General Eng Throttle Lever Position:1", "percent"},
-		{"Flaps Position", "Flaps Handle Percent", "percent"},
-		{"Gear Position", "Gear Handle Position", "bool"},
-		{"Autopilot Master", "Autopilot Master", "bool"},
-		{"Autopilot Altitude Lock", "Autopilot Altitude Lock", "bool"},
-		{"Autopilot Heading Lock", "Autopilot Heading Lock", "bool"},
-		{"Autopilot Airspeed Hold", "Autopilot Airspeed Hold", "bool"},
-		{"Autopilot Altitude Hold Value", "Autopilot Altitude Hold Var", "feet"},
-		{"Autopilot Heading Hold Value", "Autopilot Heading Lock Dir", "degrees"},
-		{"Autopilot Airspeed Hold Value", "Autopilot Airspeed Hold Var", "knots"},
-		{"Mixture Position", "General Eng Mixture Lever Position:1", "percent"},
-		{"Propeller Position", "General Eng Prop Lever Position:1", "percent"},
-	}
-
-	for _, variable := range writableVars {
-		if err := fdm.AddVariableWithWritable(variable.name, variable.simVar, variable.units, true); err != nil {
-			return err
-		}
-	}
-
 	return nil
 }
